drude

Drude model for the dielectric function of good (governed by free electrons) metals

drude(wavelength = 633, p = c(1e+16, 1e+14, 1), omega = 2 * pi * 1e+09 *
  299792458/wavelength, omega_p = p[1], gamma_p = p[2],
  epsilon_inf = p[3], ...)

Arguments

wavelength

wavelength in nm

p

vector of 3 parameters

omega

angular frequency in rad/s

omega_p

plasma frequency in rad/s

gamma_p

damping constant, in rad/s

epsilon_inf

background dielectric function

...

not used

Value

a data.frame with wavelength in nm and complex dielectric function

Details

a bacground contribution eps_inf is assumed for the core electrons